Step 2/n of refactoring `iree/compiler/Conversion/*`.
Move some canonicalizations that existed in `Utils/*` folder into
`Transforms/*` folder.
