+++ iree/compiler/Conversion/LinalgToLLVMGPU/VectorizationPass.cpp
@@ -8,9 +8,9 @@
 #include "iree/compiler/Conversion/Passes.h"
 #include "iree/compiler/Conversion/Transforms/Transforms.h"
 #include "iree/compiler/Conversion/Utils/MarkerUtils.h"
-#include "iree/compiler/Conversion/Utils/TransformUtils.h"
 #include "iree/compiler/Conversion/Utils/Utils.h"
 #include "mlir/Conversion/StandardToSPIRV/StandardToSPIRV.h"
+#include "mlir/Conversion/VectorToSCF/VectorToSCF.h"
 #include "mlir/Dialect/Linalg/Transforms/Hoisting.h"
 #include "mlir/Dialect/MemRef/Transforms/Passes.h"
 #include "mlir/Dialect/Vector/VectorOps.h"
@@ -25,7 +25,7 @@ namespace iree_compiler {
 // Patterns for vectorization
 //====---------------------------------------------------------------------===//
 
-static void populateVectorizationPatterns(OwningRewritePatternList &patterns) {
+static void populateVectorizationPatterns(RewritePatternSet &patterns) {
   // We currently don't support vectorization of generic ops with reduction.
   // TODO(thomasraoux): Add lowering for vector.multireduce ops.
   auto filterReduction = [](Operation *op) {
@@ -68,7 +68,7 @@ static Optional<SmallVector<int64_t, 4>> getGPUNativeVectorSize(Operation *op) {
   return llvm::None;
 }
 
-static void populateVectorUnrollPatterns(OwningRewritePatternList &patterns) {
+static void populateVectorUnrollPatterns(RewritePatternSet &patterns) {
   patterns.add<vector::UnrollVectorPattern>(
       patterns.getContext(),
       vector::UnrollVectorOptions().setNativeShapeFn(getGPUNativeVectorSize));
@@ -87,7 +87,7 @@ struct LinalgToLLVMGPUVectorizationPass
 
     {
       // Step 1. Vectorize
-      OwningRewritePatternList vectorizationPatterns(context);
+      RewritePatternSet vectorizationPatterns(context);
       populateVectorizationPatterns(vectorizationPatterns);
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(vectorizationPatterns));
@@ -102,7 +102,7 @@ struct LinalgToLLVMGPUVectorizationPass
 
     {
       // Lower transfer op to canonical form.
-      OwningRewritePatternList lowerTransferOpPatterns(funcOp.getContext());
+      RewritePatternSet lowerTransferOpPatterns(funcOp.getContext());
       vector::populateVectorToVectorCanonicalizationPatterns(
           lowerTransferOpPatterns);
       vector::populateVectorToVectorTransformationPatterns(
@@ -114,12 +114,12 @@ struct LinalgToLLVMGPUVectorizationPass
 
     {
       // Step 2. Unroll the vetors to native size and canonicalize.
-      OwningRewritePatternList vectorUnrollPatterns(context);
+      RewritePatternSet vectorUnrollPatterns(context);
       populateVectorUnrollPatterns(vectorUnrollPatterns);
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(vectorUnrollPatterns));
 
-      OwningRewritePatternList canonicalizationPatterns1(funcOp.getContext());
+      RewritePatternSet canonicalizationPatterns1(funcOp.getContext());
       vector::populateVectorToVectorCanonicalizationPatterns(
           canonicalizationPatterns1);
       vector::populateVectorToVectorTransformationPatterns(
@@ -127,7 +127,7 @@ struct LinalgToLLVMGPUVectorizationPass
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(canonicalizationPatterns1));
 
-      OwningRewritePatternList canonicalizationPatterns2(funcOp.getContext());
+      RewritePatternSet canonicalizationPatterns2(funcOp.getContext());
       vector::populateVectorSlicesLoweringPatterns(canonicalizationPatterns2);
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(canonicalizationPatterns2));
