+++ iree/compiler/Conversion/LinalgToSPIRV/TileAndVectorizeInOneWorkgroupPass.cpp
@@ -18,7 +18,6 @@
 #include "iree/compiler/Conversion/Passes.h"
 #include "iree/compiler/Conversion/Transforms/Transforms.h"
 #include "iree/compiler/Conversion/Utils/MarkerUtils.h"
-#include "iree/compiler/Conversion/Utils/TransformUtils.h"
 #include "iree/compiler/Conversion/Utils/Utils.h"
 #include "iree/compiler/Dialect/HAL/IR/HALDialect.h"
 #include "iree/compiler/Dialect/HAL/IR/HALOps.h"
@@ -147,7 +146,7 @@ struct PromoteConvSubviewsPattern
 }  // namespace
 
 static void populatePromotionPatterns(MLIRContext *context,
-                                      OwningRewritePatternList &patterns) {
+                                      RewritePatternSet &patterns) {
   patterns
       .insert<PromoteMatmulSubviewsPattern,
               PromoteConvSubviewsPattern<linalg::ConvInputNHWCFilterHWCFOp>>(
@@ -200,9 +199,9 @@ struct TileMatmulSubgroupPattern
 }  // namespace
 
 /// Patterns for second level tiling to target subgroups.
-static void populateTilingToSubgroupPatterns(
-    MLIRContext *context, const LaunchConfig &launchConfig,
-    OwningRewritePatternList &patterns) {
+static void populateTilingToSubgroupPatterns(MLIRContext *context,
+                                             const LaunchConfig &launchConfig,
+                                             RewritePatternSet &patterns) {
   auto getInnerTileSizeFn = [&launchConfig](
                                 OpBuilder &builder,
                                 Operation *operation) -> SmallVector<Value, 4> {
@@ -247,9 +246,9 @@ static void populateTilingToSubgroupPatterns(
 //===----------------------------------------------------------------------===//
 
 /// Patterns for third level tiling to target invocations.
-static void populateTilingToInvocationPatterns(
-    MLIRContext *context, const LaunchConfig &launchConfig,
-    OwningRewritePatternList &patterns) {
+static void populateTilingToInvocationPatterns(MLIRContext *context,
+                                               const LaunchConfig &launchConfig,
+                                               RewritePatternSet &patterns) {
   linalg::TileSizeComputationFunction getInnerTileSizeFn =
       [&launchConfig](OpBuilder &builder, Operation *operation) {
         ArrayRef<int64_t> tileSizes = launchConfig.getTileSizes(operation, 2);
@@ -333,7 +332,7 @@ static Optional<std::pair<AffineExpr, AffineExpr>> getThreadRange(
 
 static void populateVectorizationPatterns(MLIRContext *context,
                                           const LaunchConfig &launchConfig,
-                                          OwningRewritePatternList &patterns) {
+                                          RewritePatternSet &patterns) {
   linalg::insertVectorizationPatterns<linalg::FillOp, linalg::GenericOp,
                                       linalg::ContractionOpInterface>(
       patterns, linalg::LinalgVectorizationOptions(),
@@ -346,7 +345,7 @@ static void populateVectorizationPatterns(MLIRContext *context,
 //====---------------------------------------------------------------------===//
 
 static void populateVectorUnrollPatterns(MLIRContext *context,
-                                         OwningRewritePatternList &patterns) {
+                                         RewritePatternSet &patterns) {
   patterns.insert<vector::UnrollVectorPattern>(
       context,
       vector::UnrollVectorOptions().setNativeShapeFn(getSPIRVNativeVectorSize));
@@ -419,13 +418,13 @@ static void applyVectorTransformation(FuncOp funcOp) {
       targetEnv.allows(spirv::Extension::SPV_NV_cooperative_matrix);
   {
     {
-      OwningRewritePatternList vectorUnrollPatterns(funcOp.getContext());
+      RewritePatternSet vectorUnrollPatterns(funcOp.getContext());
       populateVectorUnrollPatterns(funcOp.getContext(), vectorUnrollPatterns);
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(vectorUnrollPatterns));
     }
     {
-      OwningRewritePatternList canonicalizationPatterns1(funcOp.getContext());
+      RewritePatternSet canonicalizationPatterns1(funcOp.getContext());
 
       vector::populateVectorToVectorTransformationPatterns(
           canonicalizationPatterns1);
@@ -435,7 +434,7 @@ static void applyVectorTransformation(FuncOp funcOp) {
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(canonicalizationPatterns1));
 
-      OwningRewritePatternList canonicalizationPatterns2(funcOp.getContext());
+      RewritePatternSet canonicalizationPatterns2(funcOp.getContext());
       vector::populateVectorSlicesLoweringPatterns(canonicalizationPatterns2);
       vector::populateVectorTransferLoweringPatterns(canonicalizationPatterns2);
       (void)applyPatternsAndFoldGreedily(funcOp,
@@ -447,13 +446,13 @@ static void applyVectorTransformation(FuncOp funcOp) {
         // converted to cooperative matrix matmul op.
         // TODO(thomasraoux): remove that once we support cooperative matrix
         // lowering in MLIR core.
-        OwningRewritePatternList combineTransposePatterns(funcOp.getContext());
+        RewritePatternSet combineTransposePatterns(funcOp.getContext());
         combineTransposePatterns.add<CombineContractTranspose>(
             funcOp.getContext());
         (void)applyPatternsAndFoldGreedily(funcOp,
                                            std::move(combineTransposePatterns));
       } else {
-        OwningRewritePatternList contractLoweringPatterns(funcOp.getContext());
+        RewritePatternSet contractLoweringPatterns(funcOp.getContext());
         vector::populateVectorContractLoweringPatterns(
             contractLoweringPatterns,
             vector::VectorTransformsOptions().setVectorTransformsOptions(
@@ -485,7 +484,7 @@ static void applyVectorTransformation(FuncOp funcOp) {
 //====---------------------------------------------------------------------===//
 
 static void populateTilingConvFilterPatterns(
-    MLIRContext *context, OwningRewritePatternList &patterns,
+    MLIRContext *context, RewritePatternSet &patterns,
     const LaunchConfig &launchConfig,
     linalg::LinalgTransformationFilter marker) {
   auto getTileSizeFn = [&launchConfig](OpBuilder &builder, Operation *op) {
@@ -589,10 +588,16 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
       // The promotion patterns are put separate from the tiling patterns to
       // make sure that the allocated scratchspace memory is constant sizes
       // which requires some folding to trigger.
-      OwningRewritePatternList promotionPatterns(&getContext());
+      RewritePatternSet promotionPatterns(&getContext());
       populatePromotionPatterns(context, promotionPatterns);
       (void)applyPatternsAndFoldGreedily(funcOp, std::move(promotionPatterns));
-      applyCanonicalizationPatternsForTiling(context, funcOp);
+
+      RewritePatternSet promotionCanonicalizationPatterns =
+          linalg::getLinalgTilingCanonicalizationPatterns(context);
+      populateAffineMinCanonicalizationPattern(
+          promotionCanonicalizationPatterns);
+      (void)applyPatternsAndFoldGreedily(
+          funcOp, std::move(promotionCanonicalizationPatterns));
 
       LLVM_DEBUG({
         llvm::dbgs() << "--- After workgroup memory promotion  ---\n";
@@ -605,12 +610,18 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
     // controlled by vectorization. This is needed due to historical reasons.
     // Change the second level tiling to cyclic to loops and remove this.
     if (launchConfig.useVectorize()) {
-      OwningRewritePatternList secondLevelTilingPatterns(&getContext());
+      RewritePatternSet secondLevelTilingPatterns(&getContext());
       populateTilingToSubgroupPatterns(context, launchConfig,
                                        secondLevelTilingPatterns);
       (void)applyPatternsAndFoldGreedily(funcOp,
                                          std::move(secondLevelTilingPatterns));
-      applyCanonicalizationPatternsForTiling(context, funcOp);
+
+      RewritePatternSet secondLevelTilingCanonicalizationPatterns =
+          linalg::getLinalgTilingCanonicalizationPatterns(context);
+      populateAffineMinCanonicalizationPattern(
+          secondLevelTilingCanonicalizationPatterns);
+      (void)applyPatternsAndFoldGreedily(
+          funcOp, std::move(secondLevelTilingCanonicalizationPatterns));
       promoteSingleIterationLoops(funcOp);
 
       LLVM_DEBUG({
@@ -621,7 +632,7 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
     }
 
     {
-      OwningRewritePatternList thirdLevelTilingPatterns(&getContext());
+      RewritePatternSet thirdLevelTilingPatterns(&getContext());
       populateTilingToInvocationPatterns(context, launchConfig,
                                          thirdLevelTilingPatterns);
       (void)applyPatternsAndFoldGreedily(funcOp,
@@ -643,7 +654,12 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
                                          std::move(canoncalizationPatterns));
 
       // Perform generic canonicalization.
-      applyCanonicalizationPatternsForTiling(context, funcOp);
+      RewritePatternSet threadLevelTilingCanonicalizationPatterns =
+          linalg::getLinalgTilingCanonicalizationPatterns(context);
+      populateAffineMinCanonicalizationPattern(
+          threadLevelTilingCanonicalizationPatterns);
+      (void)applyPatternsAndFoldGreedily(
+          funcOp, std::move(threadLevelTilingCanonicalizationPatterns));
 
       LLVM_DEBUG({
         llvm::dbgs() << "--- After tiling to invocations ---\n";
@@ -653,7 +669,7 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
     }
 
     {
-      OwningRewritePatternList tilingPatterns(&getContext());
+      RewritePatternSet tilingPatterns(&getContext());
       auto marker = getLinalgMatchAndReplaceMarker(
           getConvFilterTileMarker(), getVectorizeMarker(), context);
       populateTilingConvFilterPatterns(context, tilingPatterns, launchConfig,
@@ -662,7 +678,13 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
       tilingPatterns.insert<linalg::AffineMinSCFCanonicalizationPattern>(
           context);
       (void)applyPatternsAndFoldGreedily(funcOp, std::move(tilingPatterns));
-      applyCanonicalizationPatternsForTiling(context, funcOp);
+
+      RewritePatternSet convTilingCanonicalizationPatterns =
+          linalg::getLinalgTilingCanonicalizationPatterns(context);
+      populateAffineMinCanonicalizationPattern(
+          convTilingCanonicalizationPatterns);
+      (void)applyPatternsAndFoldGreedily(
+          funcOp, std::move(convTilingCanonicalizationPatterns));
 
       LLVM_DEBUG({
         llvm::dbgs() << "--- After tiling convolution filter  ---\n";
@@ -673,7 +695,7 @@ void LinalgToSPIRVTileAndVectorizeOneWorkgroupPass::runOnOperation() {
 
     if (launchConfig.useVectorize()) {
       {
-        OwningRewritePatternList vectorizationPatterns(&getContext());
+        RewritePatternSet vectorizationPatterns(&getContext());
         populateVectorizationPatterns(context, launchConfig,
                                       vectorizationPatterns);
         populateLinalgToVectorVectorizeConvPatterns(context,
