+++ iree/compiler/Conversion/LinalgToSPIRV/VectorToGPUPass.cpp
@@ -14,8 +14,8 @@
 
 #include "iree/compiler/Conversion/PassDetail.h"
 #include "iree/compiler/Conversion/Passes.h"
+#include "iree/compiler/Conversion/Transforms/Transforms.h"
 #include "iree/compiler/Conversion/Utils/MarkerUtils.h"
-#include "iree/compiler/Conversion/Utils/TransformUtils.h"
 #include "iree/compiler/Conversion/Utils/Utils.h"
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/Support/FormatVariadic.h"
@@ -91,10 +91,10 @@ void LinalgToSPIRVConvertVectorToGPUPass::tileAndVectorizeLinalgCopy(
   }
 
   // 2. Canonicalize the IR generated by tiling.
-  OwningRewritePatternList canonicalizePatterns =
+  RewritePatternSet canonicalizePatterns =
       linalg::getLinalgTilingCanonicalizationPatterns(context);
-  canonicalizePatterns.insert<AffineMinCanonicalizationPattern,
-                              linalg::AffineMinSCFCanonicalizationPattern>(
+  populateAffineMinCanonicalizationPattern(canonicalizePatterns);
+  canonicalizePatterns.insert<linalg::AffineMinSCFCanonicalizationPattern>(
       context);
   (void)applyPatternsAndFoldGreedily(funcOp, std::move(canonicalizePatterns));
