+++ iree/compiler/Conversion/Transforms/Transforms.h
@@ -12,47 +12,14 @@
 #ifndef IREE_COMPILER_CONVERSION_COMMON_TRANSFORMS_H_
 #define IREE_COMPILER_CONVERSION_COMMON_TRANSFORMS_H_
 
-#include "mlir/Dialect/Linalg/Analysis/DependenceAnalysis.h"
 #include "mlir/Dialect/Linalg/IR/LinalgOps.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
+#include "mlir/Dialect/Vector/VectorOps.h"
 #include "mlir/IR/Operation.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
 namespace iree_compiler {
-
-/// Apply canonicalizations related to tiling to make promotion/vectorization
-/// easier.
-void applyCanonicalizationPatternsForTiling(MLIRContext *context,
-                                            Operation *op);
-
-/// Assuming that `funcOp` contains a single nested scf.for that represented the
-/// tiled+fused+distributed loops with the distribution being across workgroups,
-/// i.e.
-///
-/// scf.for ... {
-///   ...
-///   scf.for ... {
-///     ...
-///     linalg.
-///     ...
-///     linalg.
-///     ...
-///   }
-/// }
-///
-/// Returns the list of linalg operations in the functions. If there are no
-/// `scf.for` operations in the function return the linalg operations in the
-/// body of the function if it has a single basic block. Return failure in all
-/// other cases.
-///
-/// TODO(ravishankarm) This methods also adds the "workgroup" marker to all ops
-/// within the loop. The marker is the way to tie into rest of the
-/// codegen. Refactor the downstream passes and get rid of the markers once and
-/// for all.
-LogicalResult getLinalgOps(FuncOp funcOp,
-                           SmallVectorImpl<linalg::LinalgOp> &linalgOps,
-                           SmallVectorImpl<Operation *> &tiledLoops);
 
 /// Specifies the number of workgroups to use for a particular entry point
 /// function, by updating the `worgroup_count` region in the
@@ -77,6 +44,26 @@ LogicalResult defineWorkgroupCountRegion(
 LogicalResult materializeStaticLaunchInformation(
     FuncOp funcOp, ArrayRef<int64_t> workloadPerWorkgroup);
 
+/// Return a fused vector::ContractionOp which represents a patterns such as:
+///
+/// ```mlir
+///    %c0 = vector.constant 0: ...
+///    %c = vector.contract %a, %b, %c0: ...
+///    %e = add %c, %d: ...
+/// ```
+///
+/// by:
+///
+/// ```mlir
+///    %e = vector.contract %a, %b, %d: ...
+/// ```
+///
+/// Return null if the canonicalization does not apply.
+// TODO: This should be a folding of Add into Contract in core but while they
+// live in different dialects, it is not possible without unnatural
+// dependencies.
+vector::ContractionOp canonicalizeContractionAdd(Operation *op);
+
 /// Insert patterns to perform folding of AffineMinOp by matching the pattern
 /// generated by tile and distribute. Try to fold a affine.min op by matching
 /// the following form:
@@ -99,6 +86,10 @@ using GetMinMaxExprFn =
 /// |getMinMaxFn| for some know values.
 void populateRemoveSingleIterationLoopPattern(RewritePatternSet &patterns,
                                               GetMinMaxExprFn getMinMaxFn);
+
+/// Insert pattern to fold chains of `affine.min` operations.
+// TODO: It is not clear what this pattern is doing and should be deprecated.
+void populateAffineMinCanonicalizationPattern(RewritePatternSet &patterns);
 
 }  // namespace iree_compiler
 }  // namespace mlir
