+++ iree/compiler/Conversion/Utils/BUILD
@@ -16,32 +16,19 @@ cc_library(
     name = "Utils",
     srcs = [
         "MarkerUtils.cpp",
-        "TransformUtils.cpp",
         "Utils.cpp",
     ],
     hdrs = [
         "MarkerUtils.h",
-        "TransformUtils.h",
         "Utils.h",
     ],
     deps = [
         "//iree/compiler/Dialect/Flow/IR",
         "//iree/compiler/Dialect/HAL/IR",
-        "//iree/compiler/Dialect/HAL/IR:HALDialect",
-        "//iree/compiler/Dialect/HAL/Utils",
-        "//iree/compiler/Dialect/Shape/IR",
         "@llvm-project//llvm:Support",
-        "@llvm-project//mlir:Affine",
-        "@llvm-project//mlir:Analysis",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:LinalgOps",
         "@llvm-project//mlir:LinalgTransforms",
-        "@llvm-project//mlir:Pass",
-        "@llvm-project//mlir:SCFDialect",
-        "@llvm-project//mlir:StandardOps",
         "@llvm-project//mlir:Support",
-        "@llvm-project//mlir:Transforms",
-        "@llvm-project//mlir:VectorOps",
-        "@llvm-project//mlir:VectorToSCF",
     ],
 )
