+++ iree/compiler/Conversion/Utils/Utils.h
@@ -38,6 +38,34 @@ Type getUntiledType(Value tiledView);
 /// `subtensor` op chain (for tensors).
 ArrayRef<int64_t> getUntiledShape(Value tiledView);
 
+/// Assuming that `funcOp` contains a single nested scf.for that represented the
+/// tiled+fused+distributed loops with the distribution being across workgroups,
+/// i.e.
+///
+/// scf.for ... {
+///   ...
+///   scf.for ... {
+///     ...
+///     linalg.
+///     ...
+///     linalg.
+///     ...
+///   }
+/// }
+///
+/// Returns the list of linalg operations in the functions. If there are no
+/// `scf.for` operations in the function return the linalg operations in the
+/// body of the function if it has a single basic block. Return failure in all
+/// other cases.
+///
+/// TODO(ravishankarm) This methods also adds the "workgroup" marker to all ops
+/// within the loop. The marker is the way to tie into rest of the
+/// codegen. Refactor the downstream passes and get rid of the markers once and
+/// for all.
+LogicalResult getLinalgOps(FuncOp funcOp,
+                           SmallVectorImpl<linalg::LinalgOp> &linalgOps,
+                           SmallVectorImpl<Operation *> &tiledLoops);
+
 }  // namespace iree_compiler
 }  // namespace mlir
